# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ServiceConnect API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Service {
  id: number;
  name: string;
  description: string;
  icon: string;
  active: boolean;
  createdAt: string;
}

export interface CreateService {
  name: string;
  description: string;
  icon: string;
  active?: boolean;
}

export type LeadStatus = (typeof LeadStatus)[keyof typeof LeadStatus];

export const LeadStatus = {
  new: "new",
  contacted: "contacted",
  forwarded: "forwarded",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface Lead {
  id: number;
  serviceId: number;
  serviceName: string;
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  address: string;
  description: string;
  status: LeadStatus;
  adminNotes?: string;
  assignedProviderId?: number;
  assignedProviderName?: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateLead {
  serviceId: number;
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  address: string;
  description: string;
}

export type UpdateLeadStatus =
  (typeof UpdateLeadStatus)[keyof typeof UpdateLeadStatus];

export const UpdateLeadStatus = {
  new: "new",
  contacted: "contacted",
  forwarded: "forwarded",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface UpdateLead {
  status?: UpdateLeadStatus;
  adminNotes?: string;
  assignedProviderId?: number;
}

export interface Provider {
  id: number;
  name: string;
  contactName: string;
  email: string;
  phone: string;
  serviceIds: number[];
  serviceNames: string[];
  rating?: number;
  active: boolean;
  notes?: string;
  createdAt: string;
}

export interface CreateProvider {
  name: string;
  contactName: string;
  email: string;
  phone: string;
  serviceIds: number[];
  rating?: number;
  notes?: string;
}

export interface UpdateProvider {
  name?: string;
  contactName?: string;
  email?: string;
  phone?: string;
  serviceIds?: number[];
  rating?: number;
  active?: boolean;
  notes?: string;
}