# lib/api-zod/src/generated/types/createLead.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ServiceConnect API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateLead {
  serviceId: number;
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  address: string;
  description: string;
}