# lib/api-zod/src/generated/types/createProvider.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ServiceConnect API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProvider {
  name: string;
  contactName: string;
  email: string;
  phone: string;
  serviceIds: number[];
  rating?: number;
  notes?: string;
}