# lib/api-zod/src/generated/types/lead.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ServiceConnect API
 * OpenAPI spec version: 0.1.0
 */
import type { LeadStatus } from "./leadStatus";

export interface Lead {
  id: number;
  serviceId: number;
  serviceName: string;
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  address: string;
  description: string;
  status: LeadStatus;
  adminNotes?: string;
  assignedProviderId?: number;
  assignedProviderName?: string;
  createdAt: Date;
  updatedAt: Date;
}