# lib/api-zod/src/generated/types/provider.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ServiceConnect API
 * OpenAPI spec version: 0.1.0
 */

export interface Provider {
  id: number;
  name: string;
  contactName: string;
  email: string;
  phone: string;
  serviceIds: number[];
  serviceNames: string[];
  rating?: number;
  active: boolean;
  notes?: string;
  createdAt: Date;
}