# lib/api-zod/src/generated/types/updateLead.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ServiceConnect API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateLeadStatus } from "./updateLeadStatus";

export interface UpdateLead {
  status?: UpdateLeadStatus;
  adminNotes?: string;
  assignedProviderId?: number;
}