# lib/api-zod/src/generated/types/updateProvider.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ServiceConnect API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateProvider {
  name?: string;
  contactName?: string;
  email?: string;
  phone?: string;
  serviceIds?: number[];
  rating?: number;
  active?: boolean;
  notes?: string;
}